+++ docs/CHANGELOG.md
@@ -24,6 +24,7 @@ General:
    - Fixed missing axis flip for `md2` scale and translate value imports
    - Added support for Quake1 `mdl` files
    - Added support for exporting and importing to and from `png` slices
+   - Added support for Autodesk `3ds` files
 
 VoxConvert: